// src/shared/types/IRoutes.ts
export interface IRoutes {
  LAYOUT: string;
  HOME: string;
  Service_Center: string;
  SIGNIN: string;
  SIGNIN_PASSWORD: string;
  SIGNUP: string;
  VERIFICATION_QUESTIONS: string;
  PREVIOUS_LOCATION: string;
  PASSWORD_DISPLAY: string;
  CHANGE_PASSWORD: string;
  EDIT_PROFILE: string;
  MY_APPLICATIONS: string;
  // DAMAGE_ASSESSMENT_DIALOG: string;
  CURRENT_LOCATION: string;
  PERSONAL_INFO: string;
  SETTINGS: string;
  FAMILY_INFO: string;
  DEPARTMENT_SECTION: string;
  RESET_PASSWORD: string;
  // DAMAGE_ASSESSMENT: string;
  // DOCUMENTS: string;
  // MAP: string;
  REVIEW: string;
  SUCCESS: string;
  TRACK_STATUS: string;
  ADMIN_LOGIN: string;
  ADMIN_DASHBOARD: string;
  // ADMIN_USERS: string;
  // ADMIN_APPLICATIONS: string;
  // ADMIN_CITIZENS: string;
  // ADMIN_LOCATIONS: string;
  ADMIN_LOCATION_MAP: string;
  ADMIN_RESET_PASSWORD: string;
  ADMIN_FORGOT_PASSWORD: string;
  CITIZEN_DASHBOARD: string;
  CITIZEN_FORGOT_PASSWORD: string;
  CITIZEN_RESET_PASSWORD: string;
  // ADMIN_BANKING: string;
  BANK_INFORMATION: string;
  BIOMETRIC_DATA: string;
  MY_COMPLAINTS: string;
  COMPLAINT_DETAILS: string;
  CENTRAL_DATABASE: string;
  PUBLIC_SERVICES: string;
  EMERGENCY_MANAGEMENT: string;
  SUPPORT_NETWORK: string;
  DAMAGE_ASSESSMENT: string;
}